requirements.txt
# MCP工具与大模型交互核心依赖
langchain-openai>=0.2.0
langchain-mcp-adapters==0.1.0
langgraph>=0.1.0
python-dotenv>=1.0.0

# 智能体框架依赖
langchain>=0.3.0
langchain-core>=0.3.36
langchain-community>=0.3.0

# 数据处理和工具
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
aiohttp>=3.8.0

# 配置和类型支持
pydantic>=2.0.0
typing-extensions>=4.5.0

# 异步支持
asyncio-mqtt>=0.13.0
# uvloop>=0.17.0  # 仅支持Unix系统，Windows用户请注释此行

# 报告生成功能 (可选依赖)
markdown>=3.4.0
python-docx>=0.8.11  # 用于生成Word文档
pdfkit>=1.0.0       # 用于生成PDF文档 (还需要安装wkhtmltopdf)
reportlab>=4.0.0

# 开发工具
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0